requirements.txt
pydantic-settings
fastapi
asyncpg
aiohttp
redis
psycopg2-binary
passlib[bcrypt]
pyjwt
python-dotenv
SQLAlchemy
uvicorn